Fixes wrong URL reporting
This fixes the incorrect variable reporting. Without this, if you inspect the raw variables in GTM preview mode, you'll see that `_url`, `Page URL` and `Page Path` are incorrectly reported. It reports the previous page, because when `router.afterEach` fires, I presume the internal browser data was not updated yet.