Update README Setup section - markdown-it-div
Hi, I worked with markdown-it-container and it's mandatory containerName variable is harder to handle in aplication if we want to give a custom classes to divs. I found that with markdown-it-div plugin it's working out-of the box so I recommend to change it in README, cause it's not even listed in makdownit example plugins section.